Move GoogleAnalytics and GoogleTagManager to blocks
