+++ docs/app/layout.tsx
@@ -1,16 +1,8 @@
 import * as React from 'react';
 import { Metadata } from 'next';
-import { GoogleTagManager } from 'docs-base/src/components/GoogleTagManager';
-import { BaseUiGoogleAnalytics } from 'docs-base/src/components/GoogleAnalytics';
-import * as Tooltip from '@base_ui/react/Tooltip';
-import { DemoVariantSelectorProvider } from 'docs-base/src/components/demo/DemoVariantSelectorProvider';
-import { PackageManagerSnippetProvider } from 'docs-base/src/blocks/PackageManagerSnippet/PackageManagerSnippetProvider';
+import { GoogleAnalytics } from 'docs-base/src/components/GoogleAnalytics';
+import { DocsProviders } from './DocsProviders';
 import '../src/styles/style.css';
-
-const PRODUCTION_GA =
-  process.env.DEPLOY_ENV === 'production' || process.env.DEPLOY_ENV === 'staging';
-
-const GOOGLE_ANALYTICS_ID_V4 = PRODUCTION_GA ? 'G-5NXDQLC2ZK' : 'G-XJ83JQEK7J';
 
 interface Props {
   children: React.ReactNode;
@@ -24,20 +16,12 @@ export default function Layout({ children }: Props) {
         <link rel="manifest" href="/static/manifest.json" />
         <link rel="shortcut icon" href="/static/favicon.ico" />
         <link rel="apple-touch-icon" sizes="180x180" href="/static/icons/180x180.png" />
-        <meta name="mui:productId" content="base-ui" />
-        <meta name="mui:productCategoryId" content="core" />
       </head>
       <body>
-        <Tooltip.Provider delay={350}>
-          <DemoVariantSelectorProvider defaultVariant="css-modules" defaultLanguage="ts">
-            <PackageManagerSnippetProvider defaultValue="npm">
-              {children}
-
-              <GoogleTagManager id={GOOGLE_ANALYTICS_ID_V4} />
-              <BaseUiGoogleAnalytics />
-            </PackageManagerSnippetProvider>
-          </DemoVariantSelectorProvider>
-        </Tooltip.Provider>
+        <DocsProviders>
+          {children}
+          <GoogleAnalytics />
+        </DocsProviders>
       </body>
     </html>
   );
